% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{as_df}
\alias{as_df}
\title{Coerce to a data frame}
\usage{
as_df(x, ...)
}
\arguments{
\item{x}{A rectangular object, with or without rownames}

\item{...}{Additional columns to include}
}
\value{
A data frame. Row names, if present in `x`, are included as a `.rownames` column
}
\description{
Coerce to a data frame
}

