% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_sampling}
\alias{calculate_sampling}
\alias{calculate_sampling_}
\title{Calculates sampling fractions for BAMM}
\usage{
calculate_sampling(phy, grouping, richness, ignore.tips = c())

calculate_sampling_(phy, grouping, richness, ignore.tips = c())
}
\arguments{
\item{phy}{an object of class \code{"phylo"}}

\item{grouping}{the grouping level at which to calculate sampling fraction}

\item{richness}{a data frame, with columns of groups (e.g., family, genus)
and a final column \code{richness} specifying how many species are in that group}

\item{tips}{that should be retained in the tree, but ignored for the
purposes of calculating taxon sampling.}
}
\value{
a list with two components:
\describe{
  \item{global}{a numeric vector of length 1; the global sampling fraction
  in \code{phy}}
  \item{clade}{a data frame with three columns: tip labels, the grouping,
  and the sampling fraction}
}
}
\description{
Calculates sampling fractions for BAMM
}
\seealso{
\code{\link{make_sampling_file}} to write a sampling file in BAMM's format
}

