% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{dist_2d}
\alias{dist_2d}
\title{Median distance of two groups}
\usage{
dist_2d(x, y, group)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}

\item{group}{a vector that specifies how \code{x} and \code{y} should be grouped}
}
\description{
Median distance of two groups
}

