% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fish.reliability.R
\docType{package}
\name{fish.reliability}
\alias{fish.reliability}
\alias{fish.reliability-package}
\title{fish.reliability: Functions and data for the fish reliability manuscript}
\description{
fish.reliability: Functions and data for the fish reliability manuscript
}
\section{Plotting functions}{

\link{pca_df}, \link{plot_tangent_space}
}

\section{BAMM functions}{

\link{calculate_sampling}, \link{make_sampling_file}
}

\section{data}{

\link{fish_reliability}, \link{fish_families}
}

