% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzzdata.R
\docType{data}
\name{fish_families}
\alias{fish_families}
\title{Shape landmarks for 7 fish families}
\format{A data frame with 10,360 rows and 6 variables:
\describe{
  \item{worker}{hash identifying each digitizer}
  \item{mark}{the name of the landmark in question}
  \item{x}{the x-coordinate of this landmark}
  \item{y}{the y-coordinate of this landmark}
  \item{family}{the family that this specimen belonged to}
  \item{tip}{the species digitized}
  \item{time_taken}{the amount of time taken by this worker to digitize this image}
}}
\usage{
fish_families
}
\description{
A dataset containing landmarks for 147 species across 7 fish families.
}
\keyword{datasets}

