% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzzdata.R
\docType{data}
\name{fish_phy}
\alias{fish_phy}
\title{Phylogenies for 7 fish families}
\format{A named list consisting of 7 phylo objects}
\usage{
fish_phy
}
\description{
A dataset containing landmarks for 147 species across 7 fish families.
}
\references{
Acanthuridae: Sorenson 2013, Balistoidae / Tetraodontidae: Santini 2013, Apogonidae, Chaetodontidae, Labridae: Cowman & Bellwood 2011. Scarinae (grafted into Labridae tree): Choat et al 2012. Pomacentridae: Frederich et al 2013
}
\keyword{datasets}

