% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{format_for_gpagen}
\alias{format_for_gpagen}
\alias{format_for_gpagen_}
\title{Formats a data frame for \code{\link[geomorph]{gpagen}}}
\usage{
format_for_gpagen(df, ...)

format_for_gpagen_(df, ..., .dots)
}
\arguments{
\item{df}{a data frame. must include columns "x" and "y"}

\item{...}{columns to include as labels}

\item{.dots}{used to work around nonstandard evaluation}
}
\value{
A 3D matrix
}
\description{
Formats a data frame for \code{\link[geomorph]{gpagen}}
}

