% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{htext}
\alias{htext}
\alias{vtext}
\title{Plot horizontal or vertical text}
\usage{
htext(str)

vtext(str)
}
\arguments{
\item{str}{string to plot}
}
\description{
Plot horizontal or vertical text
}

