% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{mad_2d}
\alias{mad_2d}
\title{2D median absolute deviation}
\usage{
mad_2d(x, y, constant = 1.4826)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector, of the same length as \code{x}.}

\item{constant}{scale factor. See 'Details' in \code{\link[stats]{mad}}.}
}
\description{
Compute the median absolute deviation for two dimensional data.
}

