% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{make_sampling_file}
\alias{make_sampling_file}
\title{Writes sampling fractions for BAMM}
\usage{
make_sampling_file(sampling, file = "")
}
\arguments{
\item{sampling}{a list with two components, from \code{\link{make_sampling}}}

\item{file}{A \link[base]{connection}, or a character string naming the file to print to. If "" (the default), prints to the standard output connection, the console unless redirected by \code{\link[base]{sink}}.}
}
\value{
None (invisible \code{NULL})
}
\description{
Writes sampling fractions for BAMM
}

