% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{my.procD.lm}
\alias{my.procD.lm}
\title{Hack the shit out of geomorph's slow procD.lm function}
\usage{
my.procD.lm(f1, iter = 999, RRPP = FALSE, int.first = FALSE,
  verbose = FALSE)
}
\description{
Hack the shit out of geomorph's slow procD.lm function
}

