% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{pca_df}
\alias{pca_df}
\title{Do PCA on Procrustes-aligned specimens}
\usage{
pca_df(coords, rename_rownames = TRUE)
}
\arguments{
\item{coords}{An array (p x k x n) containing landmark coordinates for a set of aligned specimens, possibly from \link{geomorph::gpagen}.}

\item{rename_rownames}{Rename and remove the rownames? If TRUE (the default), converts rownames to a column in the data frame called `id`. If a character string, instead calls that column by that string. If FALSE, preserves the rownames as-is.}
}
\value{
an object of class `tbl_df`, `tbl`, and `data.frame`, with columns PC1 to PC...
}
\description{
Do PCA on Procrustes-aligned specimens
}

