% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_rtt}
\alias{plot_rtt}
\title{Plot BAMM rate through time}
\usage{
plot_rtt(rtt, shifts = list(), xlim = NULL, ylim = NULL, labels = TRUE)
}
\arguments{
\item{rtt}{Rate through tiem matrix}

\item{shifts}{list of other rate through time matrices corresponding to certain nodes}

\item{xlim}{vector of length 2, x limits of plot}

\item{ylim}{vector of length 2, y limits of plot}

\item{labels}{whether to plot axis labels}
}
\description{
Plot BAMM rate through time
}

