% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_tangent_space}
\alias{plot_tangent_space}
\title{Plot specimens in tangent space}
\usage{
plot_tangent_space(coords_df, axis1 = "PC1", axis2 = "PC2")
}
\arguments{
\item{coords_df}{A PCA data frame, possibly from \link{pca_df}. Array labels must be in the format \code{group1_group2_individual}.}

\item{axis1}{The name of the first axis to plot}

\item{axis2}{The name of the second axis to plot}
}
\value{
A \code{ggplot} object
}
\description{
This function plots a set of Procrustes-aligned specimens in tangent space along their principal axes. This is a custom replacement for \code{geomorph::plotTangentSpace} tailored for the fish reliability project.
}

