% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_tangent_space2}
\alias{plot_tangent_space2}
\title{Ungrouped plot tangent space}
\usage{
plot_tangent_space2(coords_df, axis1 = "PC1", axis2 = "PC2",
  groupcol = "family", labelmethod = list("smart.grid"))
}
\description{
XXX: Make plot tangent space take a grouped_df
}

