% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotColorRamp}
\alias{plotColorRamp}
\title{Plot BAMM color ramps}
\usage{
plotColorRamp(q)
}
\arguments{
\item{q}{BAMM plot data, from \link{BAMMtools::plot.bammdata}}
}
\description{
Plots a color ramp for a BAMM phylorate plot, using the BAMM rates KDE.
}
\author{
Pascal Title
}

