% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{substitute_template}
\alias{substitute_template}
\alias{substitute_template_}
\title{Do template substitution on a file}
\usage{
substitute_template(template, ...)

substitute_template_(template, ..., .dots)
}
\arguments{
\item{template}{string with templates that look like \code{\{template\}}}

\item{...}{substitutions to perform}

\item{.dots}{Used to work around non-standard evaluation. See
\code{vignette("nse", "dplyr")} for details}
}
\value{
a character vector
}
\description{
Given a file with strings that look like \code{\{templ\}}, replace it
}
\examples{
substitute_template("{animal} says {noise}", animal="dog", noise="bark")
}

