% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{upper_outlier}
\alias{upper_outlier}
\title{Detects an upper (more positive) outlier}
\usage{
upper_outlier(x, extreme = 2)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{the cutoff point for determining an outlier}
}
\description{
Detects an upper (more positive) outlier
}

