% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{format_for_gpagen}
\alias{format_for_gpagen}
\title{Formats a data frame for `geomorph::gpagen`}
\usage{
format_for_gpagen(df)
}
\arguments{
\item{df}{a data frame, with certain columns}
}
\value{
A 3D matrix
}
\description{
Formats a data frame for `geomorph::gpagen`
}

